+++ ingesters/alnc/src/test/java/concrete/ingesters/alnc/ALNCIngesterTest.java
@@ -1,6 +1,7 @@
 package concrete.ingesters.alnc;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -14,25 +15,30 @@ import edu.jhu.hlt.concrete.AnnotationMetadata;
 import edu.jhu.hlt.concrete.Communication;
 import edu.jhu.hlt.concrete.ingesters.alnc.ALNCIngester;
 import edu.jhu.hlt.concrete.ingesters.base.IngestException;
+import edu.jhu.hlt.concrete.serialization.CompactCommunicationSerializer;
+import edu.jhu.hlt.concrete.serialization.TarGzCompactCommunicationSerializer;
+import edu.jhu.hlt.concrete.util.ConcreteException;
 
 public class ALNCIngesterTest {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ALNCIngesterTest.class);
-  
+
   Path p = Paths.get("src/test/resources/fake.json");
-  
+  CompactCommunicationSerializer cs = new TarGzCompactCommunicationSerializer();
+
   @Test
-  public void testIterator() throws IngestException {
+  public void testIterator() throws IngestException, ConcreteException {
     try (ALNCIngester ing = new ALNCIngester(p);) {
       Iterator<Communication> iter = ing.iterator();
       int ct = 0;
       while (iter.hasNext()) {
         Communication c = iter.next();
+        cs.toBytes(c);
         LOGGER.info("ID: {}", c.getId());
         LOGGER.info("UUID: {}", c.getUuid());
         AnnotationMetadata md = c.getMetadata();
         LOGGER.info("Got md: {}", md.toString());
-        assertTrue(md.getTool().contains("concrete-ingesters-alnc"));
+        assertTrue(md.getTool().contains("ALNC"));
         LOGGER.info("Got text: {}", c.getText());
         assertEquals("news", c.getType());
         ct++;