+++ server/src/test/java/concrete/server/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright 2012-2015 Johns Hopkins University HLTCOE. All rights reserved.
+ * See LICENSE in the project root directory.
+ */
+
+/**
+ * 
+ */
+package concrete.server;