+++ util/src/main/java/edu/jhu/hlt/concrete/streams/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author max
+ *
+ */
+package edu.jhu.hlt.concrete.streams;