+++ util/src/main/java/edu/jhu/hlt/concrete/util/SuperTextSpan.java
@@ -7,22 +7,26 @@ package edu.jhu.hlt.concrete.util;
 
 import edu.jhu.hlt.concrete.Communication;
 import edu.jhu.hlt.concrete.TextSpan;
+import edu.jhu.hlt.concrete.spans.TextSpanWrapper;
 
 /**
- * A wrapper around {@link TextSpan} offering improved functionality. 
- * 
- * @author max
+ * A wrapper around {@link TextSpan} offering improved functionality.
+ *
+ * @deprecated
+ *
+ * @see TextSpanWrapper
  */
+@Deprecated
 public class SuperTextSpan {
 
   private final TextSpan ts;
   private final Communication c;
-  
+
   /**
    * Create a {@link SuperTextSpan} object.
    * The {@link Communication} object should be the one that is associated
    * with the TextSpan that was passed in.
-   *  
+   *
    * @param ts - a {@link TextSpan} object
    * @param c - a {@link Communication} object from which the {@link TextSpan} object was generated
    */
@@ -34,8 +38,8 @@ public class SuperTextSpan {
   /**
    * Get the text from a {@link TextSpan}. May throw an {@link IndexOutOfBoundsException}
    * if the {@link Communication} and {@link TextSpan} do not align properly.
-   * 
-   * @return a {@link String} with the text of the {@link TextSpan} object. 
+   *
+   * @return a {@link String} with the text of the {@link TextSpan} object.
    */
   public String getText() {
     return c